Fix : details api 500 bad request error 수정 -> hooks/useFetchDetails, Chore : root layout.tsx metadata수정, Refactor : assets 폴더 삭제, public 폴더로 대체 + import경로 절대 경로로 수정
