+++ netflix-assignment/src/app/layout.tsx
@@ -4,8 +4,11 @@ import "./globals.css";
 
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "Netflix Assignment",
+  description: "Netflix Assignment",
+  icons: {
+    icon: '/favicon.ico',
+  }
 };
 
 export default function RootLayout({
